Header cleanup: r_local.h
`r_local.h` is just a header including other headers. In order to do
some header cleanup and get rid of unnecessary includes, remove it and
find out individually for each file, which headers they need to include.

`p_local.h` is giving us trouble, and we might need to split function
definitions from it to other files in the future.